+++ src/components/LanguageSelector/LanguageSelector.tsx
@@ -6,33 +6,37 @@ import { loadSettings, saveSettings } from '../../lib/storage/storage';
 import i18n from '../../config/langConfig';
 import styles from './LanguageSelector.module.scss';
 import type { Settings } from '../../types/app';
+import t from '../../lib/lang/lang';
 
 /**
  * Allows the user to switch the active interface language.
- * @param props - The props for the LanguageSelector component.
- * @param props.onChange Callback function that is called when the language is changed.
+ * @param props The props for the LanguageSelector component.
+ * @param props.setSettings A function to update the application settings.
  * @returns The rendered LanguageSelector component.
  */
-export default function LanguageSelector({ onChange }: { onChange: () => void }) {
+export default function LanguageSelector({ setSettings }: { setSettings: React.Dispatch<React.SetStateAction<Settings>> }) {
 
     const [lang, setLang] = useState(() => loadSettings().lang);
 
     useEffect(() => {
         const settings = loadSettings();
-        if (settings.lang === lang && i18n.language === lang) {
+        if (settings.lang === lang && i18n.language === lang)
             return;
-        }
 
-        settings.lang = lang;
-        saveSettings(settings);
+        setSettings((prev) => ({ ...prev, lang }));
+        saveSettings({ ...settings, lang });
 
         i18n.changeLanguage(lang);
-        onChange();
-    }, [lang, onChange]);
+    }, [lang, setSettings]);
 
 
     return (
-        <select className={styles.languageSelector} value={lang} onChange={(e) => setLang(e.target.value as Settings['lang'])}>
+        <select
+            className={styles.languageSelector}
+            value={lang}
+            onChange={(e) => setLang(e.target.value as Settings['lang'])}
+            aria-label={t('language')}
+        >
             <option value="en">English</option>
             <option value="fr">Français</option>
         </select>